Displaced effort for F and maternal contam transfer
This has displaced effort in place for mFC forced harvest (overloading Effort matrix to store the displaced F).
Also midway adding maternal transfer of contaminants - have it in recruits and migrants.

+++ atlantis/atecology/attime.c
@@ -206,7 +206,7 @@ void Time_OYr(MSEBoxModel *bm, double thisday) {
 		DaysInQ = 91;
 	else
 		DaysInQ = 92;
-	bm->HowFar = DayInQuarter / DaysInQ;
+    bm->HowFar = ((double)(DayInQuarter)) / ((double)(DaysInQ));
 	if (QuarterOfYear < 3)
 		NextQuarterOfYear = QuarterOfYear + 1;
 	else
@@ -492,7 +492,7 @@ static void Calculate_Move_Date_Index(MSEBoxModel *bm){
             orig_index = FunctGroupArray[sp].moveEntryIndex;
             
 			/* find the length of each period */
-			delta = 364.0 / FunctGroupArray[sp].numMoveEntries;
+            delta = 364.0 / ((double)(FunctGroupArray[sp].numMoveEntries));
 
 			/* now find the index of the current season. */
 			index = floor(bm->TofY / delta);
